pyproject.toml
[project]
name = "gradeup"
version = "0.1.0"
description = ""
authors = [
    {name = "Shkuratova",email = "shkuratova.darina.v@gmail.com"}
]
requires-python = ">=3.14, <4.0"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
